Important Updates
Fixed some accessibility issues. Accidentally missed a couple alt tags for a couple photos. Also, gave the video a title attribute.  Fixed my skills page as the text was hard to read on anything iPad width and up.  Resized the whole skills box and brought the list items in tighter.